requirements.txt
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
bleach==4.1.0
blinker==1.4
Bottleneck==1.3.2
cachetools==4.2.2
cassandra-driver==3.25.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
click-default-group==1.2.2
colorama==0.4.4
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
Flask==2.0.1
geomet==0.2.1.post1
gitdb==4.0.7
GitPython==3.1.18
htmlmin==0.1.12
idna==3.2
ImageHash==4.2.1
importlib-metadata==4.8.1
importlib-resources==5.2.2
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==7.0.3
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
missingno==0.5.0
mistune==0.8.4
multimethod==1.4
mysql-connector-python==8.0.26
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.4
numpy==1.19.5
packaging==21.0
pandas==1.1.5
pandas-profiling==3.0.0
pandocfilters==1.5.0
parso==0.8.2
phik==0.12.0
pickleshare==0.7.5
Pillow==8.3.2
plotly==5.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
pyarrow==5.0.0
pycparser==2.20
pydantic==1.8.2
pydeck==0.6.2
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
pywin32==301
pywinpty==1.1.4
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.26.0
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smmap==4.0.0
streamlit==0.89.0
streamlit-pandas-profiling==0.1.2
tangled-up-in-unicode==0.1.0
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.2.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
typing-extensions==3.10.0.2
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.7
urwid==2.1.2
validators==0.18.2
valuedispatch==0.0.1
visions==0.7.1
watchdog==2.1.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
zipp==3.5.0
